🛠️refactor: edit Modal hooks convention
