+++ src/components/client/Modal/hooks/index.ts
@@ -1,2 +1,2 @@
-export { default as useHandleCloseModal } from './useHandleCloseModal';
-export { default as useKeydownModal } from './useKeydownModal';
+export { default as useHandleCloseModal } from './useHandleCloseModal/useHandleCloseModal';
+export { default as useKeydownModal } from './useKeydownModal/useKeydownModal';
