+++ src/components/client/Modal/hooks/useHandleCloseModal/useHandleCloseModal.type.ts
